refactor contracts
