+++ core/commands/store/offline/datastore_test.go
@@ -1,4 +1,4 @@
-package ds
+package upload
 
 import (
 	"testing"
