--- core/commands/store/upload/helper/utils.go
@@ -1,18 +1,0 @@
-package helper
-
-import (
-	"context"
-	coreiface "github.com/TRON-US/interface-go-btfs-core"
-	"github.com/TRON-US/interface-go-btfs-core/path"
-
-	cidlib "github.com/ipfs/go-cid"
-)
-
-func GetNodeSizeFromCid(ctx context.Context, hash cidlib.Cid, api coreiface.CoreAPI) (uint64, error) {
-	leafPath := path.IpfsPath(hash)
-	ipldNode, err := api.ResolveNode(ctx, leafPath)
-	if err != nil {
-		return 0, err
-	}
-	return ipldNode.Size()
-}
